extra: handle post increment in comparisons
This improves the handling of:
	if (x++ > 5) ...
In the original we handled the unop by itself and called set_state() which
set the states for both the true and false paths.

This needs to be handled in match_comparison().  I've introduced a flag in
struct expression so that if the unop has already been handled, then we
don't handle it twice.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>