flow: add a ->parent member to statement structs
You know how I have the big_expression_stack but eventually that didn't
work and I had to add a ->parent member to the expression struct?  This is
the same thing except with statements.

This lets you find out if you are inside an if statement for example.

I have a user for this but it's going to take a while before it's ready to
push.  For now, I'm just pushing the infrastructure.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>