ranges/ptrlist: solve a leak with a big hack (second attempt)
The truth is that I don't know why the record_rl() way of freeing all the
rl ptr lists wasn't working.  I really thought it should have worked.

Anyway, the new "fix" is to create a special ptr_list allocator for rl
pointer lists and then use the clear_ptrlist_alloc() to free it.  It seems
to solve the issue I was having with running out of memory running
smatch --info on the kernel.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>